+++ .gitignore
@@ -1,16 +1,14 @@
 # compiled output
 dist
-.rpt2_cache
+coverage
 
 # dependencies
 node_modules
 
 # misc
-typings
-.nyc_output
+.rpt2_cache
 .envrc
-npm-debug.log
-yarn-error.log
+*.log
 
 # system files
 .envrc
