+++ package.json
@@ -24,7 +24,7 @@
     "watch": "rollup --config rollup.config.ts --watch",
     "docs": "typedoc --mode modules --out docs/ --excludeNotExported --excludeExternals src/index.ts",
     "lint": "tslint --project tsconfig.json --fix",
-    "test": "nyc ava tests/*.test.js",
+    "test": "jest --config jest.config.json --coverage && codecov",
     "check": "snyk test",
     "format": "prettier --write \"src/**/*.ts\"",
     "changelog": "standard-changelog"
@@ -36,10 +36,12 @@
     "@commitlint/cli": "^8.0.0",
     "@commitlint/config-conventional": "^8.0.0",
     "@pascaliske/tsconfig": "^1.0.4",
-    "ava": "^2.2.0",
+    "@types/jest": "^24.0.15",
+    "codecov": "^3.5.0",
+    "expect": "^24.8.0",
     "husky": "^3.0.0",
+    "jest": "^24.8.0",
     "lint-staged": "^9.2.0",
-    "nyc": "^14.1.1",
     "prettier": "^1.18.2",
     "rollup": "^1.17.0",
     "rollup-plugin-clear": "^2.0.7",
@@ -47,6 +49,7 @@
     "rollup-plugin-typescript2": "^0.22.0",
     "snyk": "^1.193.2",
     "standard-changelog": "^2.0.11",
+    "ts-jest": "^24.0.2",
     "tslint": "^5.18.0",
     "tslint-config-prettier": "^1.18.0",
     "typedoc": "^0.14.2",
