+++ tsconfig.json
@@ -2,9 +2,14 @@
   "extends": "@pascaliske/tsconfig",
   "compilerOptions": {
     "declarationDir": "dist",
+    "typeRoots": [
+      "typings/",
+      "node_modules/@types"
+    ],
     "lib": [
-      "es2018",
+      "esnext",
       "dom"
     ]
-  }
+  },
+  "include": ["src/**/*.ts"]
 }
